# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is P2P Gateway Documentation.",
        "title": "Bitwyre P2P Develop",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/private/main/ads/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mainflow"
                ],
                "summary": "Get Ads List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponseSchema"
                        }
                    }
                }
            }
        },
        "/private/main/crypto/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mainflow"
                ],
                "summary": "Get Crypto List",
                "operationId": "mainflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponseSchema"
                        }
                    }
                }
            }
        },
        "/private/user/profile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "operationId": "user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponseSchema"
                        }
                    }
                }
            }
        },
        "/public/healthcheck": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "HealthCheck",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.BaseResponseSchema": {
            "type": "object",
            "properties": {
                "error": {},
                "meta": {},
                "result": {},
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        }
    }
}